Modifed test in general_purpose_linter_test (#22063)
* Modifed test in general_purpose_linter_test

* Added comments

* Made suggested changes
